+++ messages/sirji_messages/messages/actions/create_assistant.py
@@ -3,11 +3,10 @@ import textwrap
 from sirji_messages import AgentEnum, ActionEnum
 from .base import BaseMessages
 
-
-class InferMessage(BaseMessages):
+class CreateAssistantMessage(BaseMessages):
 
     def __init__(self):
-        self.action = ActionEnum.INFER.name
+        self.action = ActionEnum.CREATE_ASSISTANT.name
         self.to_agent = AgentEnum.RESEARCHER.name
 
         super().__init__()
@@ -18,10 +17,12 @@ class InferMessage(BaseMessages):
             "step": "Provide the step number here for the ongoing step if any.",
             "summary": "{{Display a concise summary to the user, describing the action using the present continuous tense.}}",
             "body": textwrap.dedent("""
-            {{query}}""")})
-
-    def description(self):
-        return "Ask questions on the trained content"
+            Provide the instructions to create the assistant here.
+            """)
+        })
     
+    def description(self):
+        return "Create Assistant"
+
     def instructions(self):
         return []