+++ sirji/vscode-extension/CHANGELOG.md
@@ -1,5 +1,15 @@
 # ChangeLog
 
+## 0.0.30
+
+### Enhancements
+
+- Improved error handling for network issues.
+- Default agents are now read from the `agents` folder within the `default` folder in the codebase if not present in the `agents` folder of Sirji Studio.
+- Deprecated the keys `rules` and `definitions` from agent YAML.
+- Introduced the code review agent.
+
+
 ## 0.0.29
 
 ### Features
@@ -8,12 +18,14 @@
 - Added an open-source test case agent.
 - Added `deepseek` as a new model provider for agents.
 - Added `anthropic` as a new model provider for agents.
+- Added OpenAI Assistants API integration by allowing agents to create, sync, and infer from a RAG (Retrieval-Augmented Generation) assistant. Added `CREATE_ASSISTANT`, `INFER` and `SYNC_CODEBASE` actions in the messaging protocol for the same.
 
 ### Bug Fixes
 
 - Fixed the issue of jumping to the last entry while reading logs.
 - Constructed the path for opening the studio using the global storage API.
 - File editing actions should give the needle text respecting the number of lines in the source code.
+- Regenerate venv in case of version change during extension activation.
 
 ## 0.0.28